DiretorController

+++ models/FIlme.cs
@@ -2,14 +2,14 @@ public class Filme
 {
 
 
-    public int Id { get; set; }
-    public string Titulo { get; set; }
+    public long Id { get; set; }
+    public string Titulo{ get; set; }
     public string Ano { get; set; }
     public string Genero { get; set; }
     public long DiretorId { get; set; }
     public Diretor Diretor { get; set; }
 
-    public Filme(string titulo)
+    public Filme (string titulo)
     {
         Titulo = titulo;
